Faction enum & Grundgerüst für Entity Klasse
Es gibt einen Faction enum, mit dem man Entitys in verschiedene Gruppen unterteilen kann

Die Entity Klasse hat die meisten nötigen Grundmethoden und Properties

- Methoden müssen angepasst werden mit dem Faction enum
- Alle CompareTags müssen mit enum comparern ausgetauscht werden
- Ability Script sollte nicht crashen, wenn keine TargetPosition existiert

+++ First Game/Assets/_Scripts/_General/GameFormula.cs
@@ -41,7 +41,7 @@ public class GF
 
 
     // Formeln zum Berechnen von Cooldowns
-    public static float CalculateCooldown(float Cooldown, float CDR)
+    public static float CalculateCooldown(float Cooldown, int CDR)
     {
         return Cooldown * (100f / (100 + CDR));
     }
